+++ src/rapidfuzz/algorithm/edit_based/__init__.py
@@ -1,4 +1,0 @@
-
-# this follows https://mail.python.org/pipermail/python-ideas/2008-October/002292.html
-# since it is not quite clear how well the support is for __init__.so
-from rapidfuzz.algorithm.edit_based._initialize import *