+++ src/rapidfuzz/distance/Jaro.py
@@ -1,0 +1,1 @@
+from jarowinkler.Jaro import similarity