+++ src/rapidfuzz/distance/JaroWinkler.py
@@ -1,0 +1,1 @@
+from jarowinkler.JaroWinkler import similarity